Rename `PatientSearchResultItemView` to `PatientSearchResultItemView_Old`
