+++ app/src/main/java/org/simple/clinic/main/TheActivityComponent.kt
@@ -63,7 +63,7 @@ import org.simple.clinic.teleconsultlog.shareprescription.TeleconsultSharePrescr
 import org.simple.clinic.teleconsultlog.success.TeleConsultSuccessScreen
 import org.simple.clinic.teleconsultlog.teleconsultrecord.screen.TeleconsultNotRecordedDialog
 import org.simple.clinic.teleconsultlog.teleconsultrecord.screen.TeleconsultRecordScreen
-import org.simple.clinic.widgets.PatientSearchResultItemView
+import org.simple.clinic.widgets.PatientSearchResultItemView_Old
 
 @Subcomponent(modules = [TheActivityModule::class])
 interface TheActivityComponent :
@@ -111,7 +111,7 @@ interface TheActivityComponent :
     ConfirmDiscardChangesDialog.Injector,
     EditMedicinesScreen.Injector,
     SyncIndicatorView.Injector,
-    PatientSearchResultItemView.Injector,
+    PatientSearchResultItemView_Old.Injector,
     AllPatientsInFacilityView.Injector,
     ShortCodeSearchResultScreen.Injector,
     SettingsScreen.Injector,
