+++ app/src/main/java/org/simple/clinic/searchresultsview/SearchResultsItemType.kt
@@ -8,7 +8,7 @@ import org.simple.clinic.databinding.ListPatientSearchHeaderBinding
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.patient.PatientSearchResult
 import org.simple.clinic.widgets.ItemAdapter
-import org.simple.clinic.widgets.PatientSearchResultItemView.PatientSearchResultViewModel
+import org.simple.clinic.widgets.PatientSearchResultItemView_Old.PatientSearchResultViewModel
 import org.simple.clinic.widgets.recyclerview.BindingViewHolder
 import java.util.UUID
