+++ app/src/main/res/layout/list_patient_search_header.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<TextView android:id="@+id/patientsearch_header"
-  xmlns:android="http://schemas.android.com/apk/res/android"
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
+  android:id="@+id/patientsearch_header"
   android:layout_width="match_parent"
   android:layout_height="wrap_content"
-  android:layout_marginBottom="@dimen/spacing_12"
   android:layout_marginStart="@dimen/spacing_16"
   android:layout_marginTop="@dimen/spacing_16"
+  android:layout_marginEnd="@dimen/spacing_16"
+  android:layout_marginBottom="@dimen/spacing_12"
   android:orientation="vertical"
   android:textAllCaps="true"
   android:textAppearance="@style/Clinic.V2.TextAppearance.Body2Tag.Grey1"
