+++ app/src/main/res/layout/view_patient_search_result_old.xml
@@ -3,7 +3,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="wrap_content"
-  tools:context="org.simple.clinic.widgets.PatientSearchResultItemView"
+  tools:context="org.simple.clinic.widgets.PatientSearchResultItemView_Old"
   tools:layout_margin="@dimen/spacing_8"
   tools:parentTag="androidx.cardview.widget.CardView">
